issue315: move get_highest_version() to client/__init__.py, and add latest_version to initial LibUpdate data
